middleware

+++ routes/web.php
@@ -16,13 +16,13 @@ Route::post('/hotel', [HotelController::class, 'store']);
 Route::put('/hotel/{id}', [HotelController::class, 'update']);
 Route::delete('/hotel/{id}', [HotelController::class, 'destroy']);
 
-Route::get('/room', [RoomController::class, 'index']);
-Route::get('/room/create', [RoomController::class, 'create']);
-Route::get('/room/{id}/edit', [RoomController::class, 'edit']);
-Route::post('/room', [RoomController::class, 'store']);
-Route::get('/room/{id}', [RoomController::class, 'show']);
-Route::put('/room/{id}', [RoomController::class, 'update']);
-Route::delete('/room/{id}', [RoomController::class, 'destroy']);
+Route::get('/room', [RoomController::class, 'index'])->middleware('role:Admin');
+Route::get('/room/create', [RoomController::class, 'create'])->middleware('role:Admin');
+Route::get('/room/{id}/edit', [RoomController::class, 'edit'])->middleware('role:Admin');
+Route::post('/room', [RoomController::class, 'store'])->middleware('role:Admin');
+Route::get('/room/{id}', [RoomController::class, 'show'])->middleware('role:Admin');
+Route::put('/room/{id}', [RoomController::class, 'update'])->middleware('role:Admin');
+Route::delete('/room/{id}', [RoomController::class, 'destroy'])->middleware('role:Admin');
 
 Route::get('/', fn() => redirect('/login'));
 Route::get('/login', [AuthController::class, 'login']);
